refactor: Purchasable 패키지 하위에 service, dto 패키지 추가
