+++ backend/queue-server/src/main/java/com/wootecam/festivals/purchasable/dto/PurchasableResponse.java
@@ -1,4 +1,4 @@
-package com.wootecam.festivals.purchasable;
+package com.wootecam.festivals.purchasable.dto;
 
 public record PurchasableResponse(boolean purchasable) {
 }
