+++ backend/queue-server/src/main/java/com/wootecam/festivals/purchasable/service/CheckPurchasableService.java
@@ -1,5 +1,6 @@
-package com.wootecam.festivals.purchasable;
+package com.wootecam.festivals.purchasable.service;
 
+import com.wootecam.festivals.purchasable.dto.PurchasableResponse;
 import java.time.LocalDateTime;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
